move AutomodsummRenderer from automodsumm.py to generate.py
